refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph_folder_loading/electron-folder-management.spec.ts into folder
