+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph_folder_loading/electron-folder-management.spec.ts
@@ -14,194 +14,9 @@
  * Test project has multiple subfolder directories to test folder operations.
  */
 
-import { test as base, expect, _electron as electron } from '@playwright/test';
-import type { ElectronApplication, Page } from '@playwright/test';
 import * as path from 'path';
-import * as fs from 'fs/promises';
-import * as os from 'os';
-import type { Core as CytoscapeCore } from 'cytoscape';
-import type { ElectronAPI } from '@/shell/electron';
-
-const PROJECT_ROOT = path.resolve(process.cwd());
-
-interface ExtendedWindow {
-    cytoscapeInstance?: CytoscapeCore;
-    electronAPI?: ElectronAPI;
-}
-
-// Extend test with Electron app
-const test = base.extend<{
-    electronApp: ElectronApplication;
-    appWindow: Page;
-    testProjectPath: string;
-    tempUserDataPath: string;
-}>({
-    // Create test project with multiple folders
-    testProjectPath: async ({}, use) => {
-        const tempDir = await fs.mkdtemp(path.join(os.tmpdir(), 'voicetree-folder-mgmt-test-'));
-
-        // Create a root markdown file (required for app to auto-load the project)
-        await fs.writeFile(
-            path.join(tempDir, 'root-node.md'),
-            '# Root Node\n\nTest node in project root.'
-        );
-
-        // Create multiple folders for testing
-        const folders = ['notes', 'docs', 'archive', 'projects'];
-        for (const folder of folders) {
-            const folderPath = path.join(tempDir, folder);
-            await fs.mkdir(folderPath, { recursive: true });
-            // Create a test file in each folder
-            await fs.writeFile(
-                path.join(folderPath, `${folder}-node.md`),
-                `# ${folder.charAt(0).toUpperCase() + folder.slice(1)} Node\n\nTest node in ${folder} folder.`
-            );
-        }
-
-        // Create a nested folder for search testing
-        const nestedPath = path.join(tempDir, 'projects', 'subproject');
-        await fs.mkdir(nestedPath, { recursive: true });
-        await fs.writeFile(
-            path.join(nestedPath, 'nested-node.md'),
-            '# Nested Node\n\nTest node in nested folder.'
-        );
-
-        await use(tempDir);
-
-        // Cleanup
-        await fs.rm(tempDir, { recursive: true, force: true });
-    },
-
-    tempUserDataPath: async ({}, use) => {
-        const tempUserDataPath = await fs.mkdtemp(path.join(os.tmpdir(), 'voicetree-folder-mgmt-userdata-'));
-        await use(tempUserDataPath);
-        await fs.rm(tempUserDataPath, { recursive: true, force: true });
-    },
-
-    electronApp: async ({ testProjectPath, tempUserDataPath }, use) => {
-        // Configure to auto-load test project - use notes folder as the lastDirectory
-        // so the app loads the vault directly without showing the project picker
-        const configPath = path.join(tempUserDataPath, 'voicetree-config.json');
-        const notesPath = path.join(testProjectPath, 'notes');
-        await fs.writeFile(configPath, JSON.stringify({
-            lastDirectory: notesPath
-        }, null, 2), 'utf8');
-        console.log('[Folder Management Test] Created config to auto-load:', notesPath);
-
-        const electronApp = await electron.launch({
-            args: [
-                path.join(PROJECT_ROOT, 'dist-electron/main/index.js'),
-                `--user-data-dir=${tempUserDataPath}`
-            ],
-            env: {
-                ...process.env,
-                NODE_ENV: 'test',
-                HEADLESS_TEST: '1',
-                MINIMIZE_TEST: '1',
-                VOICETREE_PERSIST_STATE: '1'
-            },
-            timeout: 15000
-        });
-
-        await use(electronApp);
-
-        // Graceful shutdown
-        try {
-            const window = await electronApp.firstWindow();
-            await window.evaluate(async () => {
-                const api = (window as unknown as ExtendedWindow).electronAPI;
-                if (api) {
-                    await api.main.stopFileWatching();
-                }
-            });
-            await window.waitForTimeout(300);
-        } catch {
-            console.log('Note: Could not stop file watching during cleanup');
-        }
-
-        await electronApp.close();
-    },
-
-    appWindow: async ({ electronApp, testProjectPath }, use) => {
-        const window = await electronApp.firstWindow({ timeout: 15000 });
-
-        window.on('console', msg => {
-            if (msg.type() === 'error') {
-                console.log(`[BROWSER ${msg.type()}]:`, msg.text());
-            }
-        });
-
-        window.on('pageerror', error => {
-            console.error('PAGE ERROR:', error.message);
-        });
-
-        await window.waitForLoadState('domcontentloaded');
-
-        // Check if project selection screen is shown
-        const isProjectSelection = await window.locator('text=Select a project to open').isVisible({ timeout: 3000 }).catch(() => false);
-
-        if (isProjectSelection) {
-            console.log('[appWindow] Project selection screen detected - adding test project');
-
-            const notesPath = path.join(testProjectPath, 'notes');
-
-            // Wait for electronAPI to be available
-            await window.waitForFunction(() => !!(window as unknown as ExtendedWindow).electronAPI, { timeout: 5000 });
-
-            // Save the project to make it appear in the list
-            const projectName = 'test-folder-mgmt';
-            await window.evaluate(async (params: { folderPath: string; projectName: string }) => {
-                const api = (window as ExtendedWindow).electronAPI;
-                if (!api) throw new Error('electronAPI not available');
-
-                // Create and save the project
-                const project = {
-                    id: crypto.randomUUID(),
-                    path: params.folderPath,
-                    name: params.projectName,
-                    type: 'folder' as const,
-                    lastOpened: Date.now(),
-                    voicetreeInitialized: false,
-                };
-                await api.main.saveProject(project);
-            }, { folderPath: notesPath, projectName });
-
-            console.log('[appWindow] Project saved, waiting for it to appear in list');
-
-            // Wait for the project to appear in the list
-            await window.waitForTimeout(500);
-
-            // Look for the project by path (since the UI shows project paths)
-            // The saved projects have a button containing the project name
-            const projectButton = window.locator(`button:has-text("${projectName}")`);
-            const projectVisible = await projectButton.isVisible({ timeout: 3000 }).catch(() => false);
-
-            if (projectVisible) {
-                console.log('[appWindow] Clicking on test project to select it');
-                await projectButton.click();
-            } else {
-                console.log('[appWindow] Project not found in list, reloading and trying again');
-                // Reload the page and try again
-                await window.reload();
-                await window.waitForLoadState('domcontentloaded');
-                await window.waitForTimeout(1000);
-
-                const projectButtonRetry = window.locator(`button:has-text("${projectName}")`);
-                await projectButtonRetry.click({ timeout: 5000 });
-            }
-
-            console.log('[appWindow] Project selected');
-        }
-
-        // Wait for cytoscape to be ready
-        await window.waitForFunction(() => (window as unknown as ExtendedWindow).cytoscapeInstance, { timeout: 20000 });
-
-        // Wait for graph to fully load
-        await window.waitForTimeout(1000);
-
-        await use(window);
-    }
-});
+import { test, expect } from './electron-folder-management/fixtures';
+import type { ExtendedWindow } from './electron-folder-management/types';
 
 test.describe('Folder Management E2E', () => {
     test('dropdown opens with three sections', async ({ appWindow }) => {
